app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.8.0
attrs==24.2.0
bitarray==3.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.1.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mistralai==1.5.0
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.6.2
networkx==3.3
numpy==2.1.1
openai==1.63.2
opt_einsum==3.4.0
owlready2==0.47
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
patsy==1.0.1
pgmpy==0.1.26
pillow==10.4.0
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.28.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.22.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==16.0.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.8.0
rlp==4.1.0
rpds-py==0.20.0
rsa==4.9
scikit-fuzzy==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SPARQLWrapper==2.0.0
statsmodels==0.14.4
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
topsispy==0.0.1
torch==2.2.2
tornado==6.4.1
tqdm==4.67.1
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.2
web3==7.9.0
websockets==13.1
xgboost==2.1.3
yarl==1.18.3
